MAUI now calls platform checkers
The platform checks will now be called when calling the main rootchecker
